Avoid computing covariance matricies to save time and memory
Compute the errors via two 1-d arrays (time and phase) rather than
keeping then full covariance matricies of each profile.

This works because the profiles are uniformly sampled, so we can
just keep the covariance function (essencilly the first row of the
covariance matrix), and the variances for each profile.